testing getFullName function
